requirements.txt
openai>=1.0.0
requests
python-dotenv
ipython
gradio>=4.0.0
